refactor(end-of-week): decompose god class into façade + 5 mixins
Extract EndOfWeekWorkflow methods into dedicated mixins in workflows/eow/:
- AnalysisMixin: week analysis and report loading
- EvaluationMixin: PID evaluation and monthly analysis
- AIWorkoutsMixin: prompt generation and AI provider dispatch
- UploadMixin: validation, upload, and planning JSON
- ArchiveMixin: archive and success summary

Façade reduced from 1102 to 425 LOC. All 46 tests pass unchanged.
Also fixes self.verbose AttributeError in _get_workouts_api (now uses sys.argv).
