+++ magma_cycling/workflows/end_of_week.py
@@ -43,12 +43,15 @@ Metadata:
 
 import argparse
 import sys
-from datetime import date, datetime, timedelta
+from datetime import date, timedelta
 from pathlib import Path
 
 from magma_cycling.config import get_data_config, get_week_config
-from magma_cycling.planning.audit_log import OperationStatus, OperationType, audit_log
-from magma_cycling.planning.models import WeeklyPlan
+from magma_cycling.workflows.eow.ai_workouts import AIWorkoutsMixin
+from magma_cycling.workflows.eow.analysis import AnalysisMixin
+from magma_cycling.workflows.eow.archive import ArchiveMixin
+from magma_cycling.workflows.eow.evaluation import EvaluationMixin
+from magma_cycling.workflows.eow.upload import UploadMixin
 
 
 def calculate_week_start_date(week_id: str) -> date:
@@ -132,7 +135,13 @@ def calculate_weekly_transition(reference_date: date | None = None) -> tuple[str
     return week_completed, week_next, completed_start_date, next_start_date
 
 
-class EndOfWeekWorkflow:
+class EndOfWeekWorkflow(
+    AnalysisMixin,
+    EvaluationMixin,
+    AIWorkoutsMixin,
+    UploadMixin,
+    ArchiveMixin,
+):
     """Orchestrateur workflow fin de semaine."""
 
     def __init__(
@@ -258,692 +267,6 @@ class EndOfWeekWorkflow:
 
                 traceback.print_exc()
             return False
-
-    def _step1_analyze_completed_week(self) -> bool:
-        """Step 1: Analyze completed week with weekly-analysis."""
-        print("\n" + "=" * 80)
-        print(f"📊 STEP 1/7: Analyse Semaine {self.week_completed}")
-        print("=" * 80)
-        print()
-
-        if self.dry_run:
-            print("🔍 DRY-RUN: Simulation analyse semaine")
-            self.reports = {
-                "bilan_final": "[DRY-RUN] Bilan final simulé",
-                "transition": "[DRY-RUN] Transition simulée",
-            }
-            return True
-
-        try:
-            completed_week_file = (
-                self.reports_dir
-                / self.week_completed
-                / f"bilan_final_{self.week_completed.lower()}.md"
-            )
-
-            # Check if analysis already exists
-            if completed_week_file.exists():
-                print(f"  ✅ Analyse {self.week_completed} déjà existante")
-                print(f"  📁 {completed_week_file}")
-
-                # Load existing reports
-                self._load_existing_reports()
-                return True
-            else:
-                # Analysis doesn't exist - run it automatically
-                print(f"  ⚠️  Analyse {self.week_completed} introuvable")
-                print("  🤖 Lancement automatique de weekly-analysis...")
-                print()
-
-                try:
-                    from magma_cycling.workflows.workflow_weekly import (
-                        run_weekly_analysis,
-                    )
-
-                    # Run weekly-analysis programmatically (Phase 2 - modern system)
-                    run_weekly_analysis(
-                        week=self.week_completed,
-                        start_date=self.completed_start_date,
-                        data_dir=self.data_dir,
-                        ai_analysis=False,
-                    )
-
-                    # Verify that analysis was created
-                    if completed_week_file.exists():
-                        print()
-                        print(f"  ✅ Analyse {self.week_completed} générée avec succès")
-                        print(f"  📁 {completed_week_file}")
-
-                        # Load newly created reports
-                        self._load_existing_reports()
-                        return True
-                    else:
-                        print()
-                        print("  ❌ Erreur : fichier bilan_final non créé")
-                        return False
-
-                except Exception as e:
-                    print()
-                    print(f"  ❌ Erreur lors de weekly-analysis : {e}")
-                    if "--verbose" in sys.argv:
-                        import traceback
-
-                        traceback.print_exc()
-                    return False
-
-        except Exception as e:
-            print(f"  ❌ Erreur analyse : {e}")
-            return False
-
-    def _load_existing_reports(self):
-        """Load existing reports from completed week."""
-        report_files = {
-            "bilan_final": f"bilan_final_{self.week_completed.lower()}.md",
-            "transition": f"transition_{self.week_completed.lower()}.md",
-            "workout_history": f"workout_history_{self.week_completed.lower()}.md",
-            "metrics_evolution": f"metrics_evolution_{self.week_completed.lower()}.md",
-        }
-
-        week_dir = self.reports_dir / self.week_completed
-
-        for key, filename in report_files.items():
-            filepath = week_dir / filename
-            if filepath.exists():
-                self.reports[key] = filepath.read_text(encoding="utf-8")
-            else:
-                self.reports[key] = f"[{filename} non trouvé]"
-
-    def _step1b_pid_evaluation(self) -> bool:
-        """Step 1b: PID evaluation and intelligence learning."""
-        print("\n" + "=" * 80)
-        print("🧠 STEP 1b/7: Évaluation PID & Training Intelligence")
-        print("=" * 80)
-        print()
-
-        if self.dry_run:
-            print("🔍 DRY-RUN: Simulation évaluation PID")
-            return True
-
-        try:
-            print("  ℹ️  Collecte des métriques d'entraînement...")
-            print(f"  📅 Période: {self.completed_start_date} → {self.completed_end_date}")
-            print()
-
-            # Import and run PID evaluation
-            from magma_cycling.scripts.pid_daily_evaluation import (
-                PIDDailyEvaluator,
-            )
-
-            evaluator = PIDDailyEvaluator(dry_run=False)
-
-            # Run evaluation for the completed week
-            result = evaluator.run_daily_evaluation(days_back=7)
-
-            print()
-            print("  ✅ Évaluation PID terminée")
-            print("  📊 Données sauvegardées dans ~/data/monitoring/pid_evaluation.jsonl")
-            print("  🧠 Intelligence mise à jour dans ~/data/intelligence.json")
-
-            # Display test recommendation if present
-            test_rec = result.get("test_recommendation")
-            if test_rec:
-                print()
-                print("  " + "=" * 76)
-                print(f"  🎯 RECOMMANDATION DÉTECTÉE: {test_rec['status']}")
-                print("  " + "=" * 76)
-                print(f"  💡 {test_rec['message']}")
-                print(f"  📅 {test_rec['timing']}")
-                print(f"  ⏰ Dernier test: {test_rec['weeks_since_test']:.1f} semaines")
-                print(f"  💪 TSB actuel: {test_rec['tsb']:.1f}")
-                print("  " + "=" * 76)
-
-            return True
-
-        except Exception as e:
-            print(f"  ⚠️  Erreur évaluation PID (non bloquant) : {e}")
-            if "--verbose" in sys.argv:
-                import traceback
-
-                traceback.print_exc()
-            # Non-blocking: continue workflow even if PID evaluation fails
-            return True
-
-    def _step1c_monthly_analysis_if_month_end(self) -> bool:
-        """Step 1c: Generate monthly analysis if month transition detected."""
-        # Detect month transition
-        completed_month = self.completed_start_date.strftime("%Y-%m")
-        next_month = self.next_start_date.strftime("%Y-%m")
-
-        if completed_month == next_month:
-            # No month transition - skip
-            return True
-
-        # Month transition detected!
-        print("\n" + "=" * 80)
-        print(f"📊 STEP 1c/7: Analyse Mensuelle Automatique - {completed_month}")
-        print("=" * 80)
-        print()
-        print(f"  🎯 Transition de mois détectée: {completed_month} → {next_month}")
-        print(f"  📅 Génération rapport mensuel pour {completed_month}")
-        print()
-
-        if self.dry_run:
-            print("🔍 DRY-RUN: Simulation analyse mensuelle")
-            return True
-
-        try:
-            # Import monthly analyzer
-            from magma_cycling.monthly_analysis import MonthlyAnalyzer
-
-            # Generate monthly report
-            analyzer = MonthlyAnalyzer(
-                month=completed_month, provider=self.provider, no_ai=(self.provider == "clipboard")
-            )
-
-            report = analyzer.run()
-
-            if not report:
-                print()
-                print("  ⚠️  Rapport vide - aucune donnée disponible")
-                return True
-
-            # Save report to file
-            report_file = self.reports_dir / f"monthly_report_{completed_month}.md"
-            report_file.write_text(report, encoding="utf-8")
-
-            print()
-            print("  ✅ Rapport mensuel généré et sauvegardé")
-            print(f"  📁 {report_file}")
-            print()
-
-            return True
-
-        except Exception as e:
-            print()
-            print(f"  ⚠️  Erreur génération rapport mensuel (non bloquant) : {e}")
-            if "--verbose" in sys.argv:
-                import traceback
-
-                traceback.print_exc()
-            # Non-blocking: continue workflow even if monthly analysis fails
-            return True
-
-    def _step2_generate_planning_prompt(self) -> bool:
-        """Step 2: Generate planning prompt for next week."""
-        print("\n" + "=" * 80)
-        print(f"✍️  STEP 2/7: Génération Prompt Planning {self.week_next}")
-        print("=" * 80)
-        print()
-
-        if self.dry_run:
-            print("🔍 DRY-RUN: Simulation génération prompt")
-            self.planning_prompt = "[DRY-RUN] Prompt simulé"
-            return True
-
-        try:
-            print("  ℹ️  Pour générer le prompt, utilisez:")
-            print(
-                f"     poetry run weekly-planner --week-id {self.week_next} "
-                f"--start-date {self.next_start_date.strftime('%Y-%m-%d')}"
-            )
-            print()
-            print("  💡 Le prompt sera copié dans votre clipboard")
-            print()
-
-            if not self.auto:
-                response = input("  Avez-vous déjà exécuté weekly-planner ? (o/n) : ")
-                if response.lower() != "o":
-                    print("  ⚠️  Veuillez d'abord exécuter weekly-planner")
-                    return False
-
-            # In clipboard mode, prompt is already in clipboard from weekly-planner
-            print("  ✅ Prompt de planification prêt")
-            return True
-
-        except Exception as e:
-            print(f"  ❌ Erreur génération prompt : {e}")
-            return False
-
-    def _step3_get_ai_workouts(self) -> bool:
-        """Step 3: Get workouts from AI provider."""
-        print("\n" + "=" * 80)
-        print(f"🤖 STEP 3/7: Génération Workouts via {self.provider.upper()}")
-        print("=" * 80)
-        print()
-
-        if self.dry_run:
-            print("🔍 DRY-RUN: Simulation génération workouts")
-            self.workouts_content = "[DRY-RUN] Workouts simulés"
-            return True
-
-        if self.provider == "clipboard":
-            return self._get_workouts_clipboard()
-        elif self.provider == "claude_api":
-            return self._get_workouts_claude_api()
-        elif self.provider == "mistral_api":
-            return self._get_workouts_mistral_api()
-        else:
-            print(f"  ❌ Provider non supporté : {self.provider}")
-            return False
-
-    def _get_workouts_clipboard(self) -> bool:
-        """Get workouts from clipboard (manual copy-paste workflow)."""
-        print("  📋 Mode CLIPBOARD - Workflow manuel")
-        print()
-        print("  1. Collez le prompt dans votre assistant IA (Claude, Mistral, etc.)")
-        print("  2. Attendez la génération des 7 workouts")
-        print("  3. Copiez la réponse COMPLÈTE dans votre clipboard")
-        print()
-
-        # Save to file for traceability
-        workouts_file = self.planning_dir / f"{self.week_next}_workouts.txt"
-        print("  4. Sauvegardez dans un fichier:")
-        print(f"     pbpaste > {workouts_file}")
-        print()
-
-        if not self.auto:
-            input("  Appuyez sur ENTRÉE une fois les workouts sauvegardés...")
-
-        # Check if file exists
-        if workouts_file.exists():
-            self.workouts_content = workouts_file.read_text(encoding="utf-8")
-            self.workouts_file = workouts_file
-            print(f"  ✅ Workouts chargés : {len(self.workouts_content)} caractères")
-            return True
-        else:
-            print(f"  ❌ Fichier workouts introuvable : {workouts_file}")
-            return False
-
-    def _get_workouts_api(self, provider_name: str) -> bool:
-        """Get workouts from any AI API provider (AI-agnostic).
-
-        Uses AIProviderFactory for provider-agnostic implementation.
-
-        Args:
-            provider_name: Provider name (claude_api, mistral_api, openai, ollama)
-
-        Returns:
-            True if workouts generated successfully
-        """
-        print(f"  🤖 Mode {provider_name.upper()} - Génération automatique")
-        print()
-
-        try:
-            # Import dependencies
-            from magma_cycling.ai_providers.factory import AIProviderFactory
-            from magma_cycling.config import get_ai_config
-            from magma_cycling.weekly_planner import WeeklyPlanner
-
-            # Get provider config from centralized AI configuration
-            config = get_ai_config().get_provider_config(provider_name)
-
-            # Validate provider config
-            is_valid, message = AIProviderFactory.validate_provider_config(provider_name, config)
-            if not is_valid:
-                print(f"  ❌ Configuration invalide : {message}")
-                return False
-
-            print(f"  ✅ {message}")
-            print()
-
-            # Create analyzer via factory (AI-agnostic)
-            analyzer = AIProviderFactory.create(provider_name, config)
-            print(f"  ℹ️  Provider: {analyzer.get_provider_info()['provider']}")
-            if "model" in analyzer.get_provider_info():
-                print(f"  ℹ️  Modèle: {analyzer.get_provider_info()['model']}")
-            print()
-
-            # Step 1: Generate prompt with WeeklyPlanner
-            print("  📝 Génération du prompt de planification...")
-            planner = WeeklyPlanner(
-                week_number=self.week_next, start_date=self.next_start_date, project_root=Path.cwd()
-            )
-
-            # Collect metrics and context
-            planner.current_metrics = planner.collect_current_metrics()
-            planner.previous_week_bilan = planner.load_previous_week_bilan()
-            planner.context_files = planner.load_context_files()
-
-            # Generate full prompt
-            prompt = planner.generate_planning_prompt()
-            print(f"  ✅ Prompt généré ({len(prompt)} caractères)")
-            print()
-
-            # Step 2: Call AI API (provider-agnostic)
-            print(f"  🤖 Appel {provider_name.upper()} pour génération workouts...")
-            print("  ⏳ Cela peut prendre 30-60 secondes...")
-            print()
-
-            workouts_response = analyzer.analyze_session(prompt)
-
-            print(f"  ✅ Workouts générés ({len(workouts_response)} caractères)")
-            print()
-
-            # Step 3: Save to file
-            workouts_file = self.planning_dir / f"{self.week_next}_workouts.txt"
-            workouts_file.write_text(workouts_response, encoding="utf-8")
-
-            self.workouts_content = workouts_response
-            self.workouts_file = workouts_file
-
-            print(f"  💾 Workouts sauvegardés : {workouts_file}")
-            print()
-
-            return True
-
-        except Exception as e:
-            print(f"  ❌ Erreur {provider_name.upper()} : {e}")
-            if self.verbose:
-                import traceback
-
-                traceback.print_exc()
-            return False
-
-    def _get_workouts_claude_api(self) -> bool:
-        """Get workouts from Claude API automatically."""
-        return self._get_workouts_api("claude_api")
-
-    def _get_workouts_mistral_api(self) -> bool:
-        """Get workouts from Mistral API automatically."""
-        return self._get_workouts_api("mistral_api")
-
-    def _step4_validate_workouts(self) -> bool:
-        """Step 4: Validate workouts notation with format-planning logic."""
-        print("\n" + "=" * 80)
-        print("🔍 STEP 4/7: Validation Notation Workouts")
-        print("=" * 80)
-        print()
-
-        if self.dry_run:
-            print("🔍 DRY-RUN: Simulation validation")
-            return True
-
-        try:
-            # Use upload-workouts validation (already integrated)
-            # Parse workouts to validate
-            from magma_cycling.upload_workouts import WorkoutUploader
-
-            # Create temporary uploader for validation only
-            dummy_date = datetime.now()
-            uploader = WorkoutUploader(self.week_next, dummy_date)
-
-            if self.workouts_file:
-                workouts = uploader.parse_workouts_file(self.workouts_file)
-            else:
-                print("  ❌ Pas de fichier workouts à valider")
-                return False
-
-            # Validation is already done in parse_workouts_file
-            if not workouts:
-                print("  ❌ Validation échouée - workouts incomplets ou mal formatés")
-                print("  💡 Utilisez format-planning pour corriger")
-                return False
-
-            print("  ✅ Validation réussie")
-            return True
-
-        except Exception as e:
-            print(f"  ❌ Erreur validation : {e}")
-            return False
-
-    def _step5_upload_workouts(self) -> bool:
-        """Step 5: Upload workouts to Intervals.icu."""
-        print("\n" + "=" * 80)
-        print("📤 STEP 5/7: Upload Workouts vers Intervals.icu")
-        print("=" * 80)
-        print()
-
-        if self.dry_run:
-            print("🔍 DRY-RUN: Simulation upload")
-            return True
-
-        if not self.workouts_file:
-            print("  ❌ Pas de fichier workouts à uploader")
-            return False
-
-        try:
-            print(
-                f"  📅 Semaine : {self.week_next} "
-                f"({self.next_start_date.strftime('%d/%m/%Y')} → "
-                f"{self.next_end_date.strftime('%d/%m/%Y')})"
-            )
-            print()
-
-            if self.auto:
-                # Auto mode: call upload programmatically
-                print("  🤖 Mode automatique - Upload programmatique")
-                print(f"  📁 Fichier : {self.workouts_file}")
-                print()
-
-                from magma_cycling.upload_workouts import WorkoutUploader
-
-                uploader = WorkoutUploader(self.week_next, self.next_start_date)
-                workouts = uploader.parse_workouts_file(self.workouts_file)
-
-                if not workouts:
-                    print("  ❌ Erreur parsing workouts")
-                    return False
-
-                print(f"  📊 {len(workouts)} workouts à uploader")
-                print()
-
-                # Upload all workouts
-                result = uploader.upload_all(workouts, dry_run=False)
-
-                print()
-                print(f"  ✅ Upload réussi : {result['success']}/{result['total']}")
-                if result["errors"]:
-                    print(f"  ⚠️  Erreurs : {len(result['errors'])}")
-                    for error in result["errors"]:
-                        print(f"     • {error}")
-
-                return result["success"] == result["total"]
-            else:
-                # Manual mode: prompt user to upload manually
-                print("  ⚠️  Upload RÉEL vers Intervals.icu")
-                print(f"  📁 Fichier : {self.workouts_file}")
-                print()
-                response = input("  Confirmer upload ? (o/n) : ")
-                if response.lower() != "o":
-                    print("  ⚠️  Upload annulé")
-                    return False
-
-                print()
-                print("  💡 Exécutez manuellement:")
-                print(
-                    f"     poetry run upload-workouts --week-id {self.week_next} "
-                    f"--start-date {self.next_start_date.strftime('%Y-%m-%d')} "
-                    f"--file {self.workouts_file}"
-                )
-                print()
-
-                response = input("  Upload effectué avec succès ? (o/n) : ")
-                return response.lower() == "o"
-
-        except Exception as e:
-            print(f"  ❌ Erreur upload : {e}")
-            return False
-
-    def _step5b_save_planning_json(self):
-        """Step 5b: Save planning JSON with real uploaded workout data."""
-        print("\n" + "=" * 80)
-        print("💾 STEP 5b/7: Sauvegarde Planning JSON")
-        print("=" * 80)
-        print()
-
-        try:
-            if not self.workouts_file:
-                print("  ⚠️  Pas de fichier workouts - JSON template déjà créé")
-                return
-
-            # Import dependencies
-            import re
-
-            from magma_cycling.api.intervals_client import IntervalsClient
-            from magma_cycling.config import get_intervals_config
-
-            # Parse workouts file to extract metadata
-            print("  📄 Parsing fichier workouts...")
-            workouts_text = self.workouts_file.read_text(encoding="utf-8")
-
-            workout_metadata = {}
-            for match in re.finditer(
-                r"=== WORKOUT (S\d+-\d+-\w+-\w+-V\d+) ===\s*\n\s*(.+?)\((\d+)min, (\d+) TSS\)",
-                workouts_text,
-            ):
-                workout_id = match.group(1)
-                name_desc = match.group(2).strip()
-                duration = int(match.group(3))
-                tss = int(match.group(4))
-                workout_metadata[workout_id] = {
-                    "name_full": name_desc,
-                    "duration_min": duration,
-                    "tss_planned": tss,
-                }
-
-            print(f"     ✓ {len(workout_metadata)} workouts parsés")
-
-            # Get uploaded events from Intervals.icu
-            print("  🌐 Récupération events Intervals.icu...")
-            config = get_intervals_config()
-            client = IntervalsClient(athlete_id=config.athlete_id, api_key=config.api_key)
-
-            events = client.get_events(
-                oldest=self.next_start_date.isoformat(),
-                newest=self.next_end_date.isoformat(),
-            )
-
-            workouts = [
-                e
-                for e in events
-                if e.get("category") == "WORKOUT" and e.get("name", "").startswith(self.week_next)
-            ]
-            workouts_sorted = sorted(workouts, key=lambda x: x.get("start_date_local", ""))
-
-            print(f"     ✓ {len(workouts_sorted)} workouts uploadés")
-
-            # Build planning data structure
-            planning_data = {
-                "week_id": self.week_next,
-                "start_date": self.next_start_date.isoformat(),
-                "end_date": self.next_end_date.isoformat(),
-                "created_at": datetime.now().isoformat(),
-                "last_updated": datetime.now().isoformat(),
-                "version": 1,
-                "athlete_id": config.athlete_id,
-                "tss_target": 0,
-                "planned_sessions": [],
-            }
-
-            tss_total = 0
-
-            for workout in workouts_sorted:
-                workout_name = workout.get("name", "")
-                # Parse: S079-01-END-RepriseDouceLundi-V001
-                match = re.match(r"(S\d+-\d+)-(\w+)-(.+?)-(V\d+)", workout_name)
-                if not match:
-                    print(f"     ⚠️  Format invalide: {workout_name}")
-                    continue
-
-                session_id = match.group(1)  # S079-01
-                workout_type = match.group(2)  # END
-                workout_name_part = match.group(3)  # RepriseDouceLundi
-                version = match.group(4)  # V001
-
-                # Get metadata from parsing
-                meta = workout_metadata.get(workout_name, {})
-                tss = meta.get("tss_planned", 0)
-                duration = meta.get("duration_min", 0)
-                description = meta.get("name_full", workout_name_part)
-
-                tss_total += tss
-
-                workout_date = workout.get("start_date_local", "").split("T")[0]
-
-                session = {
-                    "session_id": session_id,
-                    "session_date": workout_date,  # 🔧 Renamed: date → session_date
-                    "name": workout_name_part,
-                    "session_type": workout_type,  # 🔧 Renamed: type → session_type
-                    "version": version,
-                    "tss_planned": tss,
-                    "duration_min": duration,
-                    "description": description,
-                    "status": "pending",
-                    "intervals_id": workout.get("id"),
-                    "description_hash": None,
-                }
-                planning_data["planned_sessions"].append(session)
-
-            planning_data["tss_target"] = tss_total
-
-            # 🚦 CONTROL TOWER: Create planning via WeeklyPlan model
-            json_file = self.planning_dir / f"week_planning_{self.week_next}.json"
-
-            # Validate and create planning object
-            plan = WeeklyPlan(**planning_data)
-            plan.to_json(json_file)
-
-            # 📝 AUDIT LOG: Manual log for CREATE operation (no backup needed)
-            audit_log.log_operation(
-                operation=OperationType.CREATE,
-                week_id=self.week_next,
-                status=OperationStatus.SUCCESS,
-                tool="end-of-week",
-                description=f"Created planning for {self.week_next} ({len(planning_data['planned_sessions'])} sessions, {tss_total} TSS)",
-                requested_by="end-of-week-workflow",
-            )
-
-            print()
-            print(f"  ✅ Planning JSON créé: {json_file}")
-            print(f"     • {len(planning_data['planned_sessions'])} sessions")
-            print(f"     • TSS total: {tss_total}")
-
-        except Exception as e:
-            print(f"  ⚠️  Erreur sauvegarde JSON (non bloquant): {e}")
-            if "--verbose" in sys.argv:
-                import traceback
-
-                traceback.print_exc()
-            # Non-blocking: continue workflow even if JSON save fails
-
-    def _step6_archive_and_commit(self):
-        """Step 6: Archive and commit (optional)."""
-        print("\n" + "=" * 80)
-        print("📦 STEP 6/7: Archivage et Commit")
-        print("=" * 80)
-        print()
-
-        if self.dry_run:
-            print("🔍 DRY-RUN: Simulation archivage")
-            return
-
-        print("  ℹ️  Fonctionnalité en développement")
-        print("  💡 Commitez manuellement avec:")
-        print(f"     git add {self.reports_dir}/{self.week_completed}/ {self.planning_dir}/")
-        print(
-            f'     git commit -m "feat: Complete end-of-week {self.week_completed} → {self.week_next}"'
-        )
-        print()
-
-    def _print_success_summary(self):
-        """Print success summary."""
-        print("\n" + "=" * 80)
-        print("✅ WORKFLOW TERMINÉ AVEC SUCCÈS")
-        print("=" * 80)
-        print()
-        print(f"  📊 Semaine analysée   : {self.week_completed}")
-        print(f"  📅 Semaine planifiée  : {self.week_next}")
-        print(f"  🤖 Provider utilisé   : {self.provider}")
-        if self.workouts_file:
-            print(f"  📁 Fichier workouts   : {self.workouts_file}")
-        print()
-        print("  🎯 Prochaines étapes:")
-        print("     1. Vérifiez les workouts dans Intervals.icu")
-        print("     2. Ajustez si nécessaire avec workflow-coach en servo-mode")
-        print("     3. Commitez les changements (si --archive non utilisé)")
-        print()
-        print("=" * 80)
 
 
 def parse_args(args=None):
