+++ magma_cycling/workflows/eow/__init__.py
@@ -1,0 +1,1 @@
+"""EndOfWeekWorkflow mixin modules for god class decomposition."""
